requirements.txt
flask
gunicorn
python-dotenv
requests
torch
numpy
pandas
sqlalchemy
mysql-connector-python
flask-jwt-extended
